chore: extract TimeGrid to separate file
